WIP: chore: code cleanup
